iOS: leave stored in minutes (15-min-granular foundation) — Phase 1a
Foundation for 15-minute-granular leave (and, later, draggable leave). No
behavior change yet — the UI still steps whole hours; this just gives the model
the precision. CI-validated (Domain/Store + new tests).

- StoredLeave gains a precise `minutes` field (defaulted → SwiftData lightweight
  migration); legacy `hours` kept in sync (rounded) as the back-compat fallback.
  `effectiveMinutes` = minutes ?? hours*60.
- LeaveRecord carries `minutes` + a computed `hours: Double`.
- TimecardStore: `leaveMinutes(on:)` / `setLeave(on:minutes:)` are the precise
  API; `leaveHours`/`setLeave(hours:)` stay as whole-hour wrappers for the
  holiday/schedule callers.
- periodTotals `leaveByDate` is now `[String: Double]` (was Int); the leave-in-80
  and leave-fills-schedule math drop their Int casts. All five construction sites
  (Period/Day/Metrics view models, Reminders) build Double maps.
- CSV LEAVE gains a `Minutes` column; old 3-column (`Hours`-only) backups import
  via Hours×60. Round-trips fractional leave.
- LOGIC-FREEZE §3 updated (leave = minutes, 15-min granular; whole-hour default).
- Tests: fractional store round-trip, fractional CSV round-trip, fractional
  leave through periodTotals + the over-80 gate; existing leave tests updated to
  the minutes record.

Next: Phase 1b (iOS granularity setting + stepper UI), Phase 1c (PWA mirror).

+++ iOS/Platform/Reminders.swift
@@ -54,7 +54,7 @@ enum ReminderScheduler {
         let dayset = Set(period.days)
         let entries = store.allEntries()
 
-        var leaveByDate: [String: Int] = [:]
+        var leaveByDate: [String: Double] = [:]
         for l in store.allLeave() where dayset.contains(l.date) { leaveByDate[l.date] = l.hours }
 
         // Live open entry (the running clock), if any.
